requirement.txt
django==3.2.1
djangorestframework==3.12.4
coreapi==2.3.3